requirements.txt
# Core dependencies
websockets==12.0
kafka-python==2.0.2
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Data processing
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.8.0
pre-commit==3.6.0

# Crypto specific
ccxt==4.1.77
python-binance==1.0.19
